변수 타입 변경 및 클래스 이름 변경

+++ src/main/java/com/project/user/domain/application/dto/response/SummaryResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalTime;
 
 public record SummaryResponse (
 
-        Byte score,
+        int score,
         int sleepTime,
         LocalTime bedTime
 ){ }
